refactor: pages가 아닌 하위 컴포넌트의 인터페이스 컴포넌트 내부로 조정
