+++ src/hooks/api/mypages/useGameVotesQuery.ts
@@ -1,6 +1,6 @@
 import { getGameVote } from '@/api/mypages';
 import { GameVote } from '@/types/mypages';
-import { MyBalanceGameItem } from '@/types/organisms';
+import { MyBalanceGameItem } from '@/components/organisms/MyBalanceGameList/MyBalanceGameList';
 import { useInfiniteScroll } from '@/hooks/api/mypages/useInfiniteScroll';
 
 export const useGameVotesQuery = () => {
