+++ src/hooks/api/mypages/useGameWrittensQuery.ts
@@ -1,6 +1,6 @@
 import { getGameWritten } from '@/api/mypages';
 import { GameWritten } from '@/types/mypages';
-import { MyBalanceGameItem } from '@/types/organisms';
+import { MyBalanceGameItem } from '@/components/organisms/MyBalanceGameList/MyBalanceGameList';
 import { useInfiniteScroll } from '@/hooks/api/mypages/useInfiniteScroll';
 
 export const useGameWrittensQuery = () => {
