+++ src/hooks/api/mypages/useMyCommentsQuery.ts
@@ -1,6 +1,6 @@
 import { getMyComment } from '@/api/mypages';
 import { MyComment } from '@/types/mypages';
-import { InfoItem } from '@/types/organisms';
+import { InfoItem } from '@/components/organisms/InfoList/InfoList';
 import { useInfiniteScroll } from '@/hooks/api/mypages/useInfiniteScroll';
 
 export const useMyCommentsQuery = () => {
