+++ src/hooks/api/mypages/useMyWrittensQuery.ts
@@ -1,6 +1,6 @@
 import { getMyWritten } from '@/api/mypages';
 import { MyWritten } from '@/types/mypages';
-import { MyContentItem } from '@/types/organisms';
+import { MyContentItem } from '@/components/organisms/MyContentList/MyContentList';
 import { useInfiniteScroll } from '@/hooks/api/mypages/useInfiniteScroll';
 
 export const useMyWrittensQuery = () => {
